% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsave.R
\name{gsave}
\alias{gsave}
\title{Save ggplot in (WYSIWYG) Style}
\usage{
gsave() # Select your ggplot code at first and then run gsave()
}
\description{
This function run as a Rstudio Addin
and helps you save ggplot in RStudio
}
