% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcolor.R
\name{pcolor}
\alias{pcolor}
\title{Pick Color Freely in Rstudio}
\usage{
pcolor()
pcolor(5)
}
\arguments{
\item{n}{the length of the colors need to generate}
}
\description{
This function takes a numeric value or run as a Rstudio Addin
and helps your pick color freely in RStudio
}
