% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_r_files.R
\name{search_r_files}
\alias{search_r_files}
\title{Searching in Multiple .R Files}
\usage{
search_r_files() search strings in Current Working Directory
}
\description{
This function run as a function or Rstudio Addin.
This function searches for a specified string in all `.R` files within the current working directory and its subdirectories.
}
